[show changes component] added show changes
