+++ src/js/components/UploadFile/UploadedTable.js
@@ -1,7 +1,9 @@
-import React, { useState } from 'react';
+import React, { useState, useContext } from 'react';
 import PropTypes from 'prop-types';
+import { ThemeContext } from 'styled-components';
 import { FormNextLink, Validate, CircleAlert } from 'grommet-icons';
 
+import { defaultProps } from '../../default-props';
 import { Box } from '../Box';
 import { Button } from '../Button';
 import { Text } from '../Text';
@@ -11,8 +13,8 @@ import { TableCell } from '../TableCell';
 import { TableHeader } from '../TableHeader';
 import { TableRow } from '../TableRow';
 import { closeConfirmAlert } from '../ModalPopUp';
-import { CheckBox } from '../CheckBox';
 import { Tooltip } from '../Tooltip';
+import { ShowChanges } from './ShowChanges';
 import {
   modifiedValInRule,
   isAnyFailInRule,
@@ -20,9 +22,9 @@ import {
 } from '../../utils';
 
 const UploadedTable = ({ tableData }) => {
+  const { upload } = useContext(ThemeContext) || defaultProps.theme;
   const { headers, rules } = tableData;
   const [isShowModified, setIsShowModified] = useState(true);
-  const onChange = event => setIsShowModified(event.target.checked);
 
   const getStatusColor = (showModified, isModified, result) => {
     if (showModified && isModified && result.status === 'OK') return '#8ed09c';
@@ -32,15 +34,12 @@ const UploadedTable = ({ tableData }) => {
 
   return (
     <Box>
-      <Box align="end" margin={{ bottom: 'medium' }}>
-        <CheckBox
-          label="Show Changed"
-          checked={isShowModified}
-          onChange={onChange}
-        />
-      </Box>
+      <ShowChanges
+        isShowModified={isShowModified}
+        setIsShowModified={setIsShowModified}
+      />
       <Box>
-        <Box overflow={{ vertical: 'auto' }} height="medium">
+        <Box {...(upload && upload.table && upload.table.container)}>
           <Table>
             <TableHeader>
               <TableRow>
